moved sites.jsx to area-management.jsx
